Edited index, moved some of it to db.php
friend_request_block_User_frontEnd is a file containing the code to implement the friend request and block user code, but since we don't have a way to see other user's yet that code does not do anything. It needs to be mapped to a html frontend.

Other than that I moved the creation of the connection variable in index.php to its own file so that it is seperate so that other files and "include("db.php") to check the connection. We dont need to do that but its an optional change.