Add files via upload
File that makes so there is page in the sidebar that lists the user's friends and also lists other users that belong to the same groups of the user. This is done so that the user can send friend requests to those potential friends. This allows users to have more friends.